修复空value进入异常分支问题

+++ interfaces/kits/cj/src/common_event_defination.h
@@ -49,9 +49,9 @@ struct CCommonEventData {
 };
 
 namespace OHOS::CommonEventManager {
-    constexpr int8_t NO_ERROR = 0;
-    constexpr int8_t ERR_CES_FAILED = 1;
-    constexpr int8_t ERR_NO_MEMORY = -2;
+    constexpr int32_t NO_ERROR = 0;
+    constexpr int32_t ERR_CES_FAILED = 1;
+    constexpr int32_t ERR_NO_MEMORY = -2;
 
     char *MallocCString(const std::string &origin);
     char *MallocCString(const std::string &origin, int32_t &code);
